pyproject.toml
[project]
name = "cv-ninja"
version = "0.1.0"
description = "Computer Vision annotation format converter and toolkit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "jupyter>=1.1.1",
    "pillow>=12.0.0",
    "pip>=25.3",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
cv-ninja = "cv_ninja.cli.main:cli"
